Fixed another serious Android BS
Fixed the CastToValue method not always swapping a period for a comma. This led to balance change not always being saved properly or at all.

+++ ViewModels/QuickCalculatorViewModel.cs
@@ -29,9 +29,9 @@ namespace FarmOrganizer.ViewModels
         [RelayCommand]
         void TextChanged()
         {
-            float CropAmount = Utils.CastToValue(CropAmountValue);
-            float SellRate = Utils.CastToValue(SellRateValue);
-            float PureIncome = Utils.CastToValue(PureIncomeValue);
+            double CropAmount = Utils.CastToValue(CropAmountValue);
+            double SellRate = Utils.CastToValue(SellRateValue);
+            double PureIncome = Utils.CastToValue(PureIncomeValue);
 
             if (CropAmountReadOnly)
             {
